No issue: move relevant files in :mozilla-lint-rules to perf pkg.
