+++ mozilla-lint-rules/src/test/java/org/mozilla/fenix/lintrules/perf/ConstraintLayoutPerfDetectorTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.lintrules
+package org.mozilla.fenix.lintrules.perf
 
 import com.android.resources.ResourceFolderType
 import com.android.tools.lint.checks.infrastructure.TestFiles
